cloud_storage: soft enforcement of materialized segments limit
This operates on the same basis as the reader limit: if
units are never exhausted, the segments+readers are only
cleaned up on a TTL basis.

If we run out of units, then the trim function runs in
a proactive mode where it tries to make at least half the
reader limit units available again.