rpk: add ability to run managed plugins, cleanup code
This removes tryExecPlugin, which was the v1 way of running plugins.
The code has been unneeded since autocompletion was added.

This cleans up a LOT of documentation to hopefully make it a bit
clearer. We also support more plugins under existing commands / under
each other because of the new return from ListPlugins (prior commit).

There's a few pieces of code cleanup which are a result of re-reading my
old code and trying to figure out what it was even doing. Hopefully,
it's a bit clearer this time around.

Lastly, wiring in managed plugins is not too much extra work given the
prior changes and prior commits.