rpk plugin: add 20 sha characters to plugin name
This will enable quick checks for if a plugin is up to date. More
importantly, if a plugin's sha sum is calculated *after* tarring and
compressing, our un-zipped and un-tarred sha will never compare
correctly.

Now, we stuff 20 sha256 characters into the filename. This gives us a
relatively strong (one in one million) guarantee that our plugin will
compare incorrectly if the version is different.