storage: adds a gate to concat segment data source
The concat segment data source opens segments one at a time for reading,
maintaining a current handle and stream and supporting the get() method
to read from the current segment.

When the data source is closed it checks if the current handle is
initialized and if so it closes the handle, which also closes the
stream. However it is possible that a read-request is in-flight and
suspended and resumes after close exits, thus initializing a segment
handle. This segment handle will then never be closed.

This change adds a gate to the data source, which is entered on each
read request and the stop method waits for the gate to close before
closing the handle.