archival: decrement uploads remaining conditionally
There are code paths when scheduling uploads where the upload loop can
continue but no uploads are scheduled for that loop iteration. In this
case we should not decrement uploads remaining.

This change relies on the scheduled upload result optional to decide
whether to decrement the upload remaining, instead of using
should_stop::no which could provide a false positive.