archival_policy: fix off-by-1 error for timeboxed uploads
To check that the not-uploaded tail of the log contains only non-data
batches, we need to translate actual last stable offset (i.e. exclusive
of the last message). Previosly adjusted_lso was used which is
inclusive.

This bug manifested in not uploading the last batch in the log if it
contained exactly 1 message.

(cherry picked from commit c6be08b36470e46874fb2b201e8546d049feecaa)

Conflicts:
	src/v/archival/tests/ntp_archiver_test.cc