cloud_storage: fix is_data_available for truncated case
If we have segment but none of them are ahead of
the start_offset, we should not advertise to the
Kafka read path that we have data available.

This avoids hitting an assertion later in
`first_uploaded_offset` which checks that a segment
exists at the start offset in the manifest.

Fixes https://github.com/redpanda-data/redpanda/issues/7092