tests: amend retention in test_create_or_delete_topics_while_busy
This test probably originally intended to just have a low local
retention (pre 22.3 retention.bytes meant that), and didn't get
updated when we fixed retention properties.

However, the "buggy" version of the test proved useful for
reproducing an issue:
https://github.com/redpanda-data/redpanda/pull/7094

So, fix the test, but also leave the slightly odd version
with tiny global retention in there to avoid losing the coverage.