archival: make archiver no-op for unit tests
When an archiver is created inside a unit test, it co-exists with
another archiver which has been created by the redpanda app fixture.
These two archivers run in parallel and may cause multiple issues with
assertions. The http test server sees more requests than expected and
the manifest gets advanced further than expected.

This change allows replacing the archiver in fixture with a no-op
archiver. This allows the assertions in the unit test to solely depend
on the archiver started by the test.

The no-op archiver should be put into the fixture before any data is
generated by test to make sure there is no interference.

An existing test which used request masking is also converted to use
no-op archiver. The request masking API is not removed from the imposter
as it may be useful in other places.