c/leaders_table: do not apply leader updates for deleted topics
As the metadata updates and leadership updates are asynchronous we must
tolerate reordering i.e. leadership update for already deleted
partitions or for not yet created one. Leadership table should not
apply updates for deleted partitions while keep applying updates for the
one that are not yet known.

Applied updates filtering based on the last applied revision. If the
topic doesn't exists in topics table but its revision is already known
to the table then the update will be discarded as the topic was already
deleted.

Signed-off-by: Michal Maslanka <michal@redpanda.com>