cloud_storage: avoid capturing coroutine lambda in recursive_directory_walker
This one was safe because of a ss::coroutine::lambda wrapper, but we
are moving toward banning capturing coroutine lambdas entirely, so
let's simplify vetting the code by avoiding it entirely.

It's also just a more readable pattern to bundle up the tracking
state into a struct with an update method, rather than having
it live in a bunch of function scope variables captured by
a lambda.

Related: https://github.com/redpanda-data/redpanda/issues/7114