archival: Remove dead code
The branch can never be taken with the current upload logic. Previously,
every archiver was downloading the manifest from S3 before uploading
anything. Also, every upload started from the begining of the segment.
So it was possible that after leadership transfer the segment with the
same name was aready added to the manifest (with different length).

Since the addintion of archival_metadata_stm and introductin of precise
upload boundaries this is no longer happening. After leadership transfer
the first upload starts from last uploaded offset +1.