tx/compaction: fix LSO computation during bootstrap
When the state machine is still bootstrapping, not all changes may be
applied by the time housekeeping code looks to compact the segment.

This may result in an LSO for which an incomplete log state is tracked
because not all changes were applied. Particulary aborted transactions
are not fully populated by then and the subsequent self compaction sees
missing aborted ranges and ends up retaining those batches.

We noticed this behavior in a chaos job that configured a very low
compacted segment size (1MB) and repeatedly appended data to it
injecting transfer/resize operations as that happens. This resulted in
the following sequence of actions.

- partition bootstrapping
- house keeping queried for LSO, returned offset X
- applied offset Y << X (aborted transactions not up to date)
- compaction misses some aborted ranges and commits them in the log
- client sees aborted transactions.

To avoid this, we never exceed last applied offset when computing LSO.
This delays compaction until all the changes are applied and the state
machine catches up.