tests/timely_shutdown_test: pause between failures
...and restart twice as many times.

Previously this test was running a tight loop around
inject_failure, which was 99% of the time just injecting
a failure to a node that was already isolated.

The previous restart count was low enough that it would fit
into 1-2 of the 5 second isolation periods, with a decent
chance that the restarts would happen approximately
with in one period: tweak it upward to get a more consistent
behavior of at least a couple of restarts overlapping
interesting states.