storage: do not run adjacent compaction and truncation concurrently
This was unsafe. Adjacent segment compaction takes a write lock
across the range of segments it will rerwite, but truncation
does not hold a read lock continuously: it dips in and out of
read locks inside segment member functions, so compaction
can happen between the various steps in truncation.

The segment generation mechanism only protects us in certain
cases, where the segment being truncated is modifed by
compaction, and not in cases where the segment being truncated
is e.g. closed by compaction.

This is a relatively ugly approach, but:
- Correctness: we take the lock early in both functions, before
  any segment locks are held, so there should not be risk of
  deadlocks, and it is fairly easy to reason about lock order.
- Performance: truncations are relatively rare events.  If they have
  to wait for a compaction to complete, that should not impact
  availability, because a truncating log is not on the leader
  for a raft group, so it shouldn't hold up clients.

A cleaner solution will involve a rework of how locking works.

Fixes: https://github.com/redpanda-data/redpanda/issues/7118