archival: tie ntp probe lifetime to upload loop
This commit makes the lifetime of the ntp probe owned by archivers to
the upload loop. It has the effect that only the ntp leader will be
publishing metrics. Still, if the leadership change happens while
the archiver is backing off, metrics will be available from both the new
and the old leader until the back-off is done and the upload loop stops
on the previous leader.