pp/cache: add a mutex for each client
Previously, there was a chance for numerous gate closed exceptions
because the client was evicted and client::stop is called before the
first request could complete. This scenario often occurs when there is
heavy load on the cache with many unique users.

This commit addresses the problem by including a mutex to the cached
items. Then the GC process and client dispatch may use the semaphore so
GC must wait for the current request to finish before stopping the
client.