txn: fix mark_tx_ongoing to detect staleness
mark_tx_ongoing uses get_tx to get a tx it operates on.
tx_gateway_frontend uses locks so it may seem that a tx can't be mo-
difies between tx_gateway_frontend decided to mark a tx as ongoing
and actually marking it but in rare cases a leadership may bounce to
a new node which updates the tx and then the leadership may return to
the original leader just before it starts processing mark_tx_ongoing.
In those cases there is a risk of performing a dirty write. Adding
expected_term to prevent it.