archival: use no-op archiver in tests which use custom archiver
The tests which use a custom archiver first replace the fixture archiver
with a no-op object. This is achieved by passing a port to the remote
where no connections are possible, so that the no-op archiver makes no
progress and the archiver in test is solely responsible for uploads and
assertions.