cloud_storage: trim .tx and .index files together with log segment
This is to reduce the chance of situations where we might have
one segment with it's .log but no .tx, and another segment with
its .tx but no .log.  Better to entirely evict one segment than
to partially evict two.

This is not optimal: a better strategy might be to leave the tiny
.tx and .index files in place usually, and only trim them when
our total file count (rather than size) comes close to a threshold,
but that requires a more complex rework.

Related: https://github.com/redpanda-data/redpanda/issues/6411

f