archival: enforce limit on pending deletions
Normally, when deletion fails for any segment, the
manifest is not updated, as that would lead to orphaning the segments
that were not deleted in the cloud.

However, this means that the backlog for deletion grows in an unbounded
manner which can lead to bad_allocs and destabilise the system.

This commit enforces a configurable limit on the number of pending
segment deletions. If the limit is breached, the manifest gets updated
regardless of deletion failures. This means that the segments for which
deletion failed become orphaned and will need to be removed by manual
intervention.