cluster: Add batching of commands to the ...
archival_metadata_stm. In a read replica of the partition we might need
to issue multiple archival STM commands at once. In this case it's
beneficial to put all of them into the same record batch and replicate
together. This will prevent the user of the read replica from seeing the
manifest in inconsistent state (e.g. with new segments but not truncated
yet). This will also improve efficiency since all commands are
replicated once.