redpanda_installer: retry version check on failure
We were previously using `rpk redpanda admin brokers list` directly
without supplying hosts. This resulted in us unable to retry other nodes
when the targeted node was not yet a part of the controller group.

It's less noise to hit the admin endpoint directly, so this commit
updates the behavior to use the admin endpoint directly, as we commonly
do in tests elsewhere.

Fixes #7137