pp/cache: remove debounce logic
Previously, debounce logic was written into the cache in an effort to
free resources sooner-rather-than later. However, there was a race to
which core "pings" the gc timer first during cache eviction.

This commit removes the debounce logic completely. The consequence of
this is that clients may not be freed until some seconds later since gc
runs on a 10s interval.