archival: enables service stop to be called multiple times
The service stop method stops all archivers it contains and closes its
own gate. Because closing the gate and stopping archivers is not a
repeatable operation, calling stop multiple times results in assertion
errors.

In normal application code this is fine, but for unit testing where the
service has been started as part of the fixture, it is useful to be able
to stop the archivers early and start a single archiver through which
the test can make assertions. If we stop the scheduler early, the
fixture also tries to stop the scheduler when shutting down, resulting
in errors.

This change makes the service stop call repeatable by storing the state,
making it possible to stop the service early.

The downside is that we have to cater to tests in application code, but
this approach is much simpler than making many changes to unit tests to
stop the archiver, and making the service stop callable multiple times
should not have any impact on actual application code.

(cherry picked from commit 5293407b3b949f5116f660875d9dbf7ba0f3de73)