archival: adjust tests to stop scheduler early
When a unit test is using its own archiver, we stop the scheduler
service in the fixture first, before starting the http server. This way
we are sure that no archiver in fixture has been able to upload anything
and only the archiver in test is able to make any changes to state. This
makes tests and assertions more accurate.

(cherry picked from commit 2904665f05c47e78c5c5824e362836fe2234f46d)