cloud_storage: Store only materialized segments ...
in the remote_partition. The commit changes the data structure that
remote_partition uses internally. Previously, it stored an element per
log segment in the cloud. Each element could be either
materialized_segment_ptr or offloaded_segment_state. This was needed to
lookup the manifests by their kafka offsets. The high level reader was
storing an iterator to this collection and materialized segments while
traversing it.

This approach requires extra memory since we're already storing segments
in the manifest. Also, it creates some hard problems when the underlying
manifest is not immutable. We had to use an iterator wrapper to avoid
iterator invalidation between the scheduling points. With mutable
manifest we had to do extra checks on this iterator.

This commit makes collection of segments inside remote_partition to store
only materialized segments. The key of the _segments collection is now a
model::offset. This avoids a problem when two or more segments could
potentially have the same base kafka offset if some of them didn't have
data batches.

Instead of storing the iterator the remote_partition reader now uses an
api that returns a next reader. The logic that was prevoiusly living
inside the manifest synchronization procedure is now implemented there
(e.g. handling of duplicate base kafka offsets of different segments).
This method uses manifest to find segments by the kafka offset. It
returns not only a reader but also the log offset of the next segment.
The iterator can now use this offset to advance to the next segment when
the eof is reached instead of relying on the state of the
remote_segment_record_batch_reader. This avoid potential live-lock in
some cases (when the segment reader can't reach the end of the segment
file so the next offset used to create next reader belongs to the same
segment).

(cherry picked from commit 1c33e2525d71056a643945d52703edc0a5467a22)