txn: deprecate unused recovery_policy
Long time ago we had an idea to opt in relaxed consistency as a way
to overcome pathological situations without waiting for a fix but
in reality this happens to be hard to support because if we allow
the system to operate in an erroneous state the errors could only
propagate further and cause more trouble.

Dropping the rudiments of that approach.

(cherry picked from commit 69acce53a0ac310e8fb8f2d6c88a2ea5246f36fe)