archival: avoid r-value reference parameter
r-value reference parameter is error prone because it often looks as if
the caller actually moves its value, but for this to happen the receiver
must also needs to move the value into a new place.

here there is no real performance win since moving a vector is the same
as copying a pointer and a size_t.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>