r/consensus: needs_recovery is a backoff
When do_transfer_leadersip(), if a follower is still not caught up
after prepare_transfer_leadership() is done, a `timeout` was returned.
However it's not really a timeout, it's a flap (we thought recovery
was done but it's not). This commit changes it to `exponential_backoff`
so that admin API would return a 503 (plz retry) for that rather than
a 504 (we couldn't do it in time).