storage: adjusts max timestamp on batch boundary, adds tests
Changes how timestamp is calculated for end offset search when
the searched for offset lies right at the batch boundary. In
that case the timestamp returned should be the max timestamp
for the batch, instead of that of the next batch. This change
is implemented in this commit. Additionally tests are added for this
corner case.