tm_stm: introduce tm_stm cache
Transactional coordinator keeps transient state in memory and writes
it to the log only when a user commits a tx. This approach makes red
panda transactions fast but also creates a lot of opportunities for
the unknown server errors (USE). Imagine that leadership moved to a
new coordinator, in this case the in memory context is lost and the
new coordinator may only reject a tx on commit (with USE).

This commit moves code around and introduces a decoupled from tm_stm
cache to eventually let the new leader to use rpc to fetch previous
state to reconstruct the context.

The cache should be decoupled from tm_stm because the latter gets des-
troyed on partition movements but we still want to let a new coordina-
tor to query it.