c/members_backend: fixed rebalancing when more than one node was added
An incorrect condition in `members_backend` caused rebalancing not to
stop when more than one node was added to the cluster in the same time.
The condition assumend that the allocator will always move the partition
to the node that the update is currentlty being processed. This is
incorrect assumption as the allocator should be a source of truth for
partition replica placement decisions.

Signed-off-by: Michal Maslanka <michal@redpanda.com>