c/members_backend: improved reallocation stop condition
When rebalancing partitions it may happen that the partition number
isn't even or node are placed in racks so that the balancer will not be
able to achieve even number of partitions per node. Added a stop
condition based on monitoring the last improvement of unevenness error.
If balancing improvement is insignificant then the balancing is stopped.

Signed-off-by: Michal Maslanka <michal@redpanda.com>