cluster: allow seeds-driven bootstrap when empty_seed_starts_cluster=1
Previously, empty_seed_starts_cluster=1 meant that the node was
configured to use root-driven bootstrap and that it should expect some
singular node in the cluster to have an empty seeds list. This is the
default, and was mainly there for compatibility reasons. However, this
default is currently explicitly incompatible with seeds-driven
bootstrap, and thus is incompatible with having all node configurations
in the cluster being identical.

To encourage usage of seeds-driven bootstrap, and to allow homogeneous
node configs by default, this commit makes empty_seed_starts_cluster=1
compatible with seeds-driven bootstrap. An empty seed_servers list
still indicates that the configured node should be the root node, but
seeds-driven bootstrap can be used with empty_seed_starts_cluster=1 if:
- there exists no root node among any of the seeds, and
- all seeds meet the existing requirements to perform seeds-driven
  bootstrap.