(split) k8s: Gather all information from kind cluster
Due to failing PostStart hook it seems that one of the Redpanda broker is in
degraded state and AdminAPI is not responsive. The Pod log collector does not
work if pod is in `PodInitializing` state.

To find more evidence what is happening kind export logs command was added.
It should be collected by the buildkite agent and served for later
investigations.

REF
https://github.com/redpanda-data/redpanda/issues/7396