k/alter_configs: do not try to set replication factor to default
When topic property is not set in alter_configs request it is set to
defaut. We need to make an exception for replication factor to make sure
that it will not be accidentally changed to default when changing topic
configuration.

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit dbf4a8c86524b6c754e9cab5507c85729ac4483a)