tm_stm: step down on update_tx failure
Transactions in kafka protocol are stateful: the processing of the
requests depends on the previous commands executed by the same or
even different producer. It makes the situations when the replica-
tion fails with the indecisive errors such as timeout dangerous
because the true state is unknown.

At the same time we have a fundamental invariant enforcing that a
new leader start processing the request only when it has seen all
the messages written by the previous leader.

Switching tm_stm to use this invariant to handle the replication
uncertainties by doing a forced step down.