txn: update add partition & group to detect staleness
Add partition & group uses get_tx to get a tx it operates on.
tx_gateway_frontend uses locks so it may seem that a tx can't be
modifies between tx_gateway_frontend decided to update a tx and
actually updating it but in rare cases a leadership may bounce to
a new node which updates the tx and then the leadership may return
to the original leader just before it starts processing the update.
In those cases there is a risk of performing a dirty write. Adding
expected_term to prevent it.

(cherry picked from commit 1614e2e78b192bf6383d2e7f7c72e20149ba640e)