c/rm_stm: fixed returning non initialized offset as LSO
In Kafka LSO denotes the first offset that is greater from the one for
which all the transactions are resolved. Given the definition LSO is
never negative.

Fixed `rm_stm` by applying the 'next offset' rule in the situation in
which LSO is not initialized.

Fixes: #7438

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit a56d0d992d2366ab193f6778ffedd08301ed0362)