rpk: avoid decommissioning a node if its in maintenance mode
If we perform maintenance on a node, and then
decommision the same node, we will end up with
a node being permanently in maintenance mode.

To avoid this, we are introducing this preventive
action in rpk while it's implemented in redpanda
directly.

See #5601

(cherry picked from commit 6631d38a4fd7c2f02cf4093880dcab29ef56d75f)