tests: remove unused TxProducer helper
This overlapped with other tools (kgo-verifier, and
the Java transactional verifier), and its use resulted
in relatively weak tests due to its low rate of
traffic generation (or slow tests if you tried to use
larger volumes of traffic).

The volume of traffic in transaction tests is important,
to get coverage of the ways that a transaction can e.g.
span a segment boundary, or span the seam between
S3 data and local disk data.