features: include original version in feature table snapshot
On upgrades, this means that existing clusters will report
an invalid_version original version until the next write
to the feature table, that will happen when the whole cluster
is upgraded and the logical version updates.

So effectively this new field will only be meaningfully readable
once the cluster is fully upgraded.