c/members_balancer: calculate unevenness error separately per domain
The unevenness error is domain specific as it depends from the number of
replicas. To be able to determine the improvement accurately we must
calculate the error and improvement after rebalancing for each domain
separately. If the improvement from calculated reallocations is to small
they are dropped and the update is claimed finished.

Signed-off-by: Michal Maslanka <michal@redpanda.com>