c/members_backend: refined condition checking min improvement
The new replicas placement optimization stop condition is based on evenness error.

When there is no improvement of unevenness error after requesting partition
rebalancing the iteration stops.

Error is calculated as:

                                   N
                                  ___
                                  ╲    |R - r |
                                  ╱    |     n|
                                  ‾‾‾
                                 n = 0
                          e    = ──────────────
                           raw        T ⋅ N

                              T
                          R = ─
                              N

Where:

T - total number or replicas in cluster
R - requested replicas per node
N - number of nodes in the cluster
r_n - number of replicas on the node n

then the error is normalized to be in range [0,1].
To do this we calculate the maximum error:
                                             N
                                            ___
                                            ╲
                                 (T - R) +  ╱    R
                                            ‾‾‾
                                           n = 1
                          e    = ─────────────────
                           max         T ⋅ N

normalized error is equal to:

                                e
                          e = ────
                              e
                               max

Since the improvement depends on the number of replicas moved in a single batch
we need to calculate the improvement that would moving a single partition
replica introduce. Since single move requested by a balancer moves partition
from node A->B (where A has more than requested number of replicas per
and B should have less than requested number of replicas) we can calculate the
error improvement introduced by single move as:

                                    2
                          e     = ─────
                           step   N ⋅ T

Single step error must be normalized as well. We stop the movement if there in
no improvement from at least 10 percents of moves requested in a batch if ten
percent would result in 0 requested moves we require at least one
move to improve the unevenness.

Signed-off-by: Michal Maslanka <michal@redpanda.com>