cluster: re-enable partition_moving_test
This test was disabled for a long time.  Much has changed.  It
passed in a loop for a couple of hours on my desktop: even if it turns
out to still fail sometimes in CI, and we disable it again, that'll
be useful information.

Related: https://github.com/redpanda-data/redpanda/issues/2909