transactions: preserve monotonicity of LSO
Kafka clients expect that the LSO monotonically increases. In a normal
operation of the partition this holds true but if there is a change
in leadership with transactions, it may take time for the state machine
to catch up before returning the correct LSO. In that window the broker
is expected to return a retryable error code.

(cherry picked from commit 7aa02b5d7943fad6b35a54073de793c54bc30e21)