cluster: ignore removed nodes for maintenance commands
CONFLICT: replace node_id() with idx()

Issue #7370 is fixed on dev by a wide refactor that isn't ideal for
backporting. This commit is a lighter-weight fix for the issue.

This commit ignores removed nodes when considering whether nodes can be
placed into maintenance mode, matching the behavior on dev.

Fixes #7370