c/members_backend: include removed broker when building brokers set
Previously members table when asked for a broker reference returned both
the removed broker instances as well as the ones that are current
members. Fixed bug in members backend that may lead to situation in
which removed node wasn't returned while building initial partition
configuration. This will not be needed when we will not keep brokers in
raft configuration.

Signed-off-by: Michal Maslanka <michal@redpanda.com>