cluster: pass partition 0 a copy of its topic config
This enables the archiver to write the topic manifest
for topics using cloud storage.

This is a bit of a hack: a more general solution in future
might be to move the topic manifest writing into a separate
service that deals exclusively in metadata uploads, and perhaps
run that on controller leader instead of having each topic's
partition 0 leader to their topic manifest uploads.