tests: use default logging for archival/cloud_storage
We previously explicitly set these to `debug` to have more logs and make
it easier to debug issues. Now that CI runs with trace logs by default,
that intention is lost, and we end up with less logging.

This commit returns the archival/cloud_storage subsystems to the
default, and leaves the remaining explicitly set subsystems, as they
tend to be much chattier.