txn: make commit_tx respect rejected requests
txn coordinator dispatches commit_tx after it already acknowledges
a transaction to the user so in theory the commit can't fail and the
coordinator may simply retry on any error. However the coordinator
doesn't access the log when it retries so there may be a new coor-
dinator which already committed the tx and it doesn't make sense to
continue the retries. There and in similar cases rm_stm uses
request_rejected error code to stop retrying.