cloud_storage: Remove assertion from delete_object
In AWS S3 the DeleteObject call never returns NoSuchKey error. Because
of that the code in 'cloud_storage::remote' doesn't expect this error
code and triggers an assertion. But in GCP the corresponding API call
may return NoSuchKey error. This commit removes the assertion and
converts NoSuchKey to the successful outcome. This is save because our
retention is implemented expect this behavior and may call DeleteObjects
more than once for the same object.

(cherry picked from commit ccefedbaac087443b8a900e3dbaa6810a4a622da)