redpanda: attempt leadership drain during shutdown
This is helpful to users who are not using tools that
do not integrate with maintenance mode: prior to shutdown,
attempt to drain leaderships, as much as we can within 5
seconds.

This is skipped if anything in the cluster health status
indicates that we are not the only node going down, or
the rest of the cluster is unhealthy -- in these cases
we bias toward shutting down as promptly as possible.