s3: move error handling from remote to client
This is one of the most significant changes in this series of commits.
Error handling is moved from 'cloud_storage::remote' to 's3::client'.
This removes the need for the remote to know how to interpret S3 error
codes and, therefore, will allow for different client implementations in
the future.

In the new error handling flow, the client returns and 'error_outcome'
to the remote, which informs its behaviour (i.e. retry, give up, etc.).
To this end, the client catches and handles all of its exceptions (apart
from 'abort_requested_exception' which is propagated as previously) in
the new 'send_request' member function. Once caught, the exception is
translated into an error outcome and propagated to the remote via
'result<>'.