omb/perf: increase warmup time for perf run
It appears that the leadership balancer is kicking in during the
benchmark. This patch increases the warmup time to 5 mins in the hopes
that the partition assignment stabilizes by then and the cluster reaches
a stable state.

(cherry picked from commit 5116e48bb1ff4c8ae4520d115d6e7e07b411ecca)