s3: add storage client abstract class
This commit introduces an abstract class which servers as an interface
for all cloud storage clients. The client pool is also modified to use
the generic type polymorphically.

Includes of the client header have been cleane up across the tree (the
s3 client header was included in a number of places where it wasn't
required). For this purpose a few common types have also moved into
's3/types.h'.