raft: respect leader priority order in transfer_leadership
...when the caller does not specify an explicit leader.

Previously, if all peers had an equivalent last_dirty point,
then whichever one had a lower ID would receive the leadership
transfer, leading to an imbalance in leaderships that the
leader balancer would have to come back and fix later.

This is the code path used in maintenance mode, and more
recently in the drain-on-shutdown path.