ducktape: extend timeout to consumer validation.
PR #7607 bumped timeout for data generation and
compaction but turns out on ARM CDT, consumption
is slow too.

Extends the timeout to consumer validation.

Note: We don't have DEBUG/TRACE logs to validate if
the consumer is making progress but given this
only fails occasionally it probably is just
very close to the existing timeout.