cluster: wait for leadership with auto create
With the newly created topics via auto creation,
the patch additionally waits for the leadership
to be assinged before handing of the control to the
caller.

Typically auto creation is used in conjunction with
a produce which immediately runs after the creation
returns the control only to see that the newly
created partitions are leader-less. This is a
common pattern when unit testing with redpanda
(as well as regular use cases) and is a bad UX.