cloud_storage_clients: handle GCS deletion 404
The GCS S3 implementation returns a NoSuchKey 404 error response when
attempting to delete an object that doesn't exist. This commit handles
this this divergence between AWS and GCS in the client by mapping the
'notfound' outcome to a successful response. A unit test was also
introduced to cover this scenario.