rpk: support reading array of structs from files
We sometimes want to read input from files. We already do this for `rpk
group seek`, and we are adding it to `rpk acl users create`.

This adds the ability to read json, yaml, or tab/space delimited files
directly into types. For json and yaml, the types need struct tags and
the files will have keyed structs.

Also updates the gh workflow to use the newer gofumpt for lang 1.19,
since the formatting is slightly different.