kafka/client/client: Don't call client::mitigate_error during connect
A network error during connection should result in a limited retry
with backoff.  The hope is that the broker will come online, but if it
doesn't, it will give up.

Calling client::mitigate_error subverts the backoff loop, resulting
in a tight (recursive) loop of reconnections without mitigating the
error.

The purpose of the change introduced in 3ca886134ec3f5e0df4b2b53f097c5d68237b4c3
was to allow pandaproxy and schema registry to handle `sasl_authentication`
errors in order to detect the need to propagate ephemeral credentials.

The correct fix is to call external_mitigate.

Signed-off-by: Ben Pope <ben@redpanda.com>